refactor(server): move services up to top level
