+++ server/test/setup.ts
@@ -4,7 +4,7 @@ const aliases = {
   "@compilerDiagnostics": __dirname + "/../src/compilerDiagnostics",
   "@analyzer": __dirname + "/../src/parser/analyzer",
   "@common": __dirname + "/../src/parser/common",
-  "@services": __dirname + "/../src/parser/services",
+  "@services": __dirname + "/../src/services",
 };
 
 moduleAlias.addAliases(aliases);
